backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate",
    "flask",
    "flask-socketio",
    "joblib",
    "numpy",
    "opencv-python",
    "transformers==4.54.1",
    "mblt-model-zoo[transformers]>=0.3.2",
    "torch>=2.6.0",
    "python-dotenv>=1.1.1",
    "maccel>=0.30.0",
]

[tool.black]
line-length = 120
